NO SCOPE
Actually, JK
Got around the string problem by changing the scope of argv[1] to be
global.

argv[1] is a constant in program memory, thus I took advantage of its
adress and assigned it to a variable that had a global scope.  aka. I
made a global pointer.

Sight edits to the tokenizer struct, I removed str rom it, the destroy
(which I added a few edits to) and the tcreate